refactor: Make more/better use of components
The ListBooks Component was renamed to MyBooks
because the name seems more appropriate.
The single book is now a component called
ShowBook, so it can be reused on Search Page.
All shelfs are now build from a Data-Structure
called shelfMeta so a new shelf could easily be
added and HTML isn't needed to get repeated.